check spelling of column names

+++ databricks/btc_daily_OHLCV_CoinAPI_ingestion.py
@@ -245,8 +245,8 @@ time_columns = [
 ]
 
 doubleType_columns = [
-    "pice_low",
-    "prince_high",
+    "price_low",
+    "price_high",
     "price_open",
     "price_close",
     "volume_traded",
@@ -326,11 +326,15 @@ def analyze_dates(df, date_column):
 # COMMAND ----------
 
 df = spark.table(TABLE_NAME).toPandas()
-df_unique = df.drop_duplicates(keep='first')
-spark_df = spark.createDataFrame(df_unique)
+duplicate_count = df.duplicated(keep='first').sum()
+print(f"number of duplicates in table: {duplicate_count}")
 
-spark.sql(f"DROP TABLE IF EXISTS {TABLE_NAME}")
-spark_df.write.saveAsTable(TABLE_NAME)
+if duplicate_count:
+    df_unique = df.drop_duplicates(keep='first')
+    spark_df = spark.createDataFrame(df_unique)
+
+    spark.sql(f"DROP TABLE IF EXISTS {TABLE_NAME}")
+    spark_df.write.saveAsTable(TABLE_NAME)
 
 # COMMAND ----------
 
@@ -364,14 +368,14 @@ else:
 # COMMAND ----------
 
 # MAGIC %md
-# MAGIC ## Looks like missing data is missing on CoinAPI.io data, not from bad code in this notebook
+# MAGIC ## Looks like missing data is missing on CoinAPI.io data..
 
 # COMMAND ----------
 
-for d in missing_dates[0:2]:
+for d in missing_dates:
     print(f"d: {d.strftime('%Y-%m-%d')}")
-    start = (d + timedelta(days=-2)).strftime("%Y-%m-%d")
-    end = (d + timedelta(days=2)).strftime("%Y-%m-%d")
+    start = (d + timedelta(days=0)).strftime("%Y-%m-%d")
+    end = (d + timedelta(days=0)).strftime("%Y-%m-%d")
     print(f"start: {start}")
     print(f"end: {end}")
     df = coinapi_request(ENDPOINT, start, end, COINAPI_API_KEY)
